server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==1.1.0
pydantic==2.5.0
openpyxl
pandas
sqlmodel==0.0.14
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
